alu/flask/requirements.txt
SQLAlchemy==1.4.41

fastapi
uvicorn
# logger

email-validator
pyfiglet 

werkzeug
flask_bcrypt

pyfiglet
elastic-apm

python-decouple
